Move wrong-altitude tests down to core (issue #1056)
- multibyte_regression_test.rs: two tests were a strict subset of
  prescan.rs's sequence_with_multibyte_after_dash_does_not_panic; the
  other three (quoted-scalar reader, caret renderer) move to
  core/src/document/tests/multibyte_tests.rs.
- default_values_test.rs: the authored/default/zero ladder is proven in
  quill/resolved.rs, which also pins resolve() == compile_data();
  test_nested_null_zero_fills_in_plate is sole coverage and stays.
- spec_conformance_probe.rs: Document::parse is a pass-through to
  decompose_with_warnings, so the seven scenarios duplicated by
  assemble_tests/card_fence_tests prove nothing extra. Keeps the \xc2\xa78 input
  caps and the normalize_document pass.
- quill/tests.rs: the fence grammar is exhaustively covered in
  card_fence_tests.rs; keep the two wiring cases, drop the six
  re-derivations.
