+++ crates/core/src/document/tests/mod.rs
@@ -8,6 +8,7 @@ mod emit_tests;
 mod ext_tests;
 mod fence_conformance_tests;
 mod lossiness_tests;
+mod multibyte_tests;
 mod number_edge_tests;
 mod seed_tests;
